Change folder structure and remove eslint
