+++ package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
     "start": "vite preview",
     "test": "jest",
@@ -18,7 +17,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.1",
-    "sass": "^1.79.4"
+    "sass": "^1.79.4",
+    "sass-embedded": "^1.79.4"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
@@ -29,14 +29,9 @@
     "@testing-library/user-event": "^14.5.2",
     "@vitejs/plugin-react": "^4.2.1",
     "babel-jest": "^29.7.0",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "sass-embedded": "^1.79.4",
-    "vite": "^5.2.0"
+    "vite": "^5.4.8"
   },
   "jest": {
     "moduleNameMapper": {
