--- src/components/Home.css
@@ -1,11 +1,0 @@
-.home-container {
-  background-color: rgba(255, 255, 255, 0.7);
-  padding: 20px;
-  border-radius: 8px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  backdrop-filter: blur(10px);
-}
-
-.home-container p {
-  color: black;
-}
