+++ src/components/Home/Home.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './Home.css';
+
+function Home() {
+  return (
+    <div className="home-container">
+      <h1>Welcome to the Todo App</h1>
+      <p>Click on the "To-Do List" above to start managing your tasks!</p>
+    </div>
+  );
+}
+
+export default Home;
