+++ src/components/tests/InputWithLabel.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import InputWithLabel from '../InputWithLabel';
+import InputWithLabel from '../Todo/InputWithLabel';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
