+++ vite.config.js
@@ -1,8 +1,16 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   base: '/',
   plugins: [react()],
-})
+  css: {
+    preprocessorOptions: {
+      scss: {
+        // Example: custom options for Sass
+        additionalData: `@import "src/styles/variables.scss";` // if you have global variables
+      },
+    },
+  },
+});
